# packages/qn-addon/openapi.yaml
openapi: 3.0.3
info:
  title: Veil Privacy Suite — QuickNode Add-On
  description: |
    End-to-end encryption, secret sharing, and ZK compression for Solana dApps.
    MEV-protect orders, manage encrypted state, and gate features by on-chain reputation.

    ## Features
    - **NaCl Box Encryption** — Curve25519-XSalsa20-Poly1305 for end-to-end encrypted messaging
    - **Shamir Secret Sharing** — M-of-N threshold schemes for multi-party key management
    - **Order Encryption** — MEV-protected swap orders for Solana DEX integrations
    - **ZK Compression** — Light Protocol integration for 99% state cost reduction
    - **Reputation Tiers** — FairScore-based feature gating with 5 tier levels
    - **Payload Serialization** — Type-safe binary encoding for on-chain data

    ## Authentication
    API endpoints use rate limiting (100 requests/minute).
    Provisioning endpoints require HTTP Basic Auth.
  version: 0.1.0
  contact:
    name: Veil Privacy Suite Support
    email: support@veil-privacy.dev
  license:
    name: MIT
servers:
  - url: http://localhost:3030
    description: Local development
  - url: https://your-addon.example.com
    description: Production
...
paths:
...
  /v1/keypair/derive:
    post:
      tags: [Crypto]
      summary: Derive keypair from seed
      description: Deterministically derives an X25519 keypair from a 32-byte seed. Same seed always produces the same keypair.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seed]
              properties:
                seed:
                  type: string
                  format: base64
                  description: 32-byte seed encoded as base64
      responses:
        '200':
          description: Derived keypair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeypairResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/encrypt:
    post:
      tags: [Crypto]
      summary: Encrypt data
      description: |
        Encrypts plaintext using NaCl box (Curve25519-XSalsa20-Poly1305).
        The `bytes` field in the response contains the combined nonce + ciphertext for on-chain storage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plaintext, recipientPublicKey, senderSecretKey, senderPublicKey]
              properties:
                plaintext:
                  type: string
                  format: base64
                  description: Data to encrypt (base64)
                recipientPublicKey:
                  type: string
                  format: base64
                  description: Recipient's X25519 public key (base64)
                senderSecretKey:
                  type: string
                  format: base64
                  description: Sender's secret key (base64)
                senderPublicKey:
                  type: string
                  format: base64
                  description: Sender's public key (base64)
      responses:
        '200':
          description: Encrypted data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncryptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/decrypt:
    post:
      tags: [Crypto]
      summary: Decrypt data
      description: Decrypts NaCl box ciphertext back to plaintext.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bytes, senderPublicKey, recipientSecretKey, recipientPublicKey]
              properties:
                bytes:
                  type: string
                  format: base64
                  description: Combined nonce + ciphertext (base64)
                senderPublicKey:
                  type: string
                  format: base64
                recipientSecretKey:
                  type: string
                  format: base64
                recipientPublicKey:
                  type: string
                  format: base64
      responses:
        '200':
          description: Decrypted plaintext
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  plaintext:
                    $ref: '#/components/schemas/EncodedBytes'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/crypto/encrypt-multiple:
    post:
      tags: [Crypto]
      summary: Encrypt for multiple recipients
      description: |
        Encrypts the same plaintext for up to 50 recipients simultaneously.
        Each recipient gets their own encrypted copy with a unique nonce.
        Useful for broadcast messaging, group notifications, or multi-party protocols.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plaintext, recipientPublicKeys, senderSecretKey, senderPublicKey]
              properties:
                plaintext:
                  type: string
                  format: base64
                recipientPublicKeys:
                  type: array
                  items:
                    type: string
                    format: base64
                  minItems: 1
                  maxItems: 50
                  description: Array of recipient X25519 public keys (base64)
                senderSecretKey:
                  type: string
                  format: base64
                senderPublicKey:
                  type: string
                  format: base64
      responses:
        '200':
          description: Encrypted data for each recipient
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  recipientCount:
                    type: integer
                  recipients:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/EncryptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/crypto/validate:
    post:
      tags: [Crypto]
      summary: Validate encrypted data structure
      description: |
        Validates that encrypted bytes have the correct NaCl box structure
        (nonce + ciphertext) without decrypting. Useful as a pre-flight check
        before attempting decryption.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bytes]
              properties:
                bytes:
                  type: string
                  format: base64
                  description: Encrypted bytes to validate
                minPlaintextSize:
                  type: integer
                  default: 1
                  description: Minimum expected plaintext size in bytes
                maxPlaintextSize:
                  type: integer
                  default: 1024
                  description: Maximum expected plaintext size in bytes
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  valid:
                    type: boolean
                    description: Whether the data has valid NaCl box structure
                  byteLength:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/crypto/key-convert:
    post:
      tags: [Crypto]
      summary: Convert between key formats
      description: |
        Converts encryption public keys between raw bytes (base64/hex) and
        Solana base58 format. Provide either `publicKey` or `base58`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                publicKey:
                  type: string
                  format: base64
                  description: Public key as base64 (converts to base58)
                base58:
                  type: string
                  description: Public key as base58 (converts to bytes)
      responses:
        '200':
          description: Converted key in both formats
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  base58:
                    type: string
                  publicKey:
                    $ref: '#/components/schemas/EncodedBytes'
        '400':
          $ref: '#/components/responses/BadRequest'

  # === THRESHOLD ENDPOINTS ===
...
  /v1/orders/encrypt:
    post:
      tags: [Orders]
      summary: Encrypt swap order
      description: |
        Encrypts a swap order payload (minOutputAmount, slippageBps, deadline)
        using NaCl box. Only the designated solver can decrypt.
        Protects order parameters from MEV bots before on-chain submission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [minOutputAmount, slippageBps, deadline, solverPublicKey, userSecretKey, userPublicKey]
              properties:
                minOutputAmount:
                  type: string
                  description: Minimum output amount (lamports/smallest unit)
                  example: "5000000"
                slippageBps:
                  type: integer
                  description: Slippage tolerance in basis points (50 = 0.5%)
                  example: 50
                deadline:
                  type: integer
                  description: Unix timestamp for order expiration
                  example: 1700000000
                solverPublicKey:
                  type: string
                  format: base64
                userSecretKey:
                  type: string
                  format: base64
                userPublicKey:
                  type: string
                  format: base64
      responses:
        '200':
          description: Encrypted order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncryptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/orders/decrypt:
    post:
      tags: [Orders]
      summary: Decrypt swap order
      description: Decrypts an encrypted swap order. Only the solver with the correct keypair can decrypt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bytes, userPublicKey, solverSecretKey, solverPublicKey]
              properties:
                bytes:
                  type: string
                  format: base64
                userPublicKey:
                  type: string
                  format: base64
                solverSecretKey:
                  type: string
                  format: base64
                solverPublicKey:
                  type: string
                  format: base64
      responses:
        '200':
          description: Decrypted order payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  payload:
                    type: object
                    properties:
                      minOutputAmount:
                        type: string
                      slippageBps:
                        type: integer
                      deadline:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/orders/validate:
    post:
      tags: [Orders]
      summary: Validate encrypted order
      description: |
        Validates that encrypted bytes have the correct structure for a swap order
        (correct size range for NaCl box + serialized order payload) without decrypting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bytes]
              properties:
                bytes:
                  type: string
                  format: base64
                  description: Encrypted order bytes to validate
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  valid:
                    type: boolean
                  byteLength:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'

  # === PAYLOAD ENDPOINTS ===
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Auth for provisioning endpoints
...
  schemas:
    EncodedBytes:
      type: object
      properties:
        base64:
          type: string
          format: base64
        hex:
          type: string
    KeypairResponse:
      type: object
      properties:
        success:
          type: boolean
        publicKey:
          $ref: '#/components/schemas/EncodedBytes'
        secretKey:
          $ref: '#/components/schemas/EncodedBytes'
    EncryptedResponse:
      type: object
      properties:
        success:
          type: boolean
        nonce:
          $ref: '#/components/schemas/EncodedBytes'
        ciphertext:
          $ref: '#/components/schemas/EncodedBytes'
        bytes:
          $ref: '#/components/schemas/EncodedBytes'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'